survey_donkey/requirements.txt
Django==4.2.5
dj-database-url==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers
djangorestframework-simplejwt==5.3.1
sendgrid==6.11.0